fix: convert all co-ordinates to the Decimal class
This was done to resolve the issue where performing the dot_product on two normalized parallel vectors whose coordinates were of the float type resulted in round-off errors. As such a number slightly greater than 1 was calculated resulting in a domain error for the get_angle() nethod. Using the Decimal class to represent co-oordiantes in the Vecotr class increased the precision of the computation to resolve the round-off errors.