feat: implement is_parallel() and is_ortho() methods
is_parallel() determines if two vectors are parallel. This is done by calculating the normalized dot product of the two vectors and verifying that the result is close to one.

is_ortho() determines whether two vectors are orthogonal. This is done by calculating the dot product of both vectors and verifying that they are close to zero.