fix: pass 'coordinates' variable of Vector object to first_nonzero_index()
The first_nonzero_index() method expects an iterable object. However, this
method was being accessed within the Line class by passing in a object of
the vector class. As a result, an exception would be raised when creating
a Line object and passing in to the constructor an argument that is a
Vector object.