fix: Use local variable for converting Decimal to Float within Vector
Whenever the __str__ built-in method was called on a Vector object the
co-ordinate attributes would be changed from Decimal to Float. This
would cause an exception to be thrown. The fix was to use a variable
local to __str__ to store the result of 'type casting' from Double to
Float.